:arrow_up: upgrade dependencies
The only dependency that caused trouble was auth0-js on the functions
side. auth0-js is actually a browser-side only library and should not
have been used in functions. It worked by chance. However auth0-js 9.7
changed the dependency it uses to make HTTP requests to a browser-only
version of superagent, which broke the Auth0 interaction in the
exchangeToken function. auth0-js was consequently replaced by the proper
Node.js Auth0 library: auth0.