:recycle: enforce typescript strict mode
Strict mode was disabled because it was not compatible with firebase-functions@1.0.0.
Now that we are using @2.0.0 I tried to enable it again. I found firebase-functions@2.0.0
to work OK but firebase-tools@5.1.1 that does not have types and so does not compile
under strict mode. However this time instead of leaving strict mode disable I tried to
find a way to make it work regardless and found one. I created our own type declarations
for this library.