:recycle: switch noImplicitAny back on (#44)
We had to switch the `noImplicitAny` TypeScript compiler option off when [upgrading to `firebase-functions@1.0.0`](https://github.com/Zenika/humeur-du-mois-2018/commit/65a61084f369915c588dc896bec62a443a67d77a#diff-dc8de0a814a0e019ed6c1deb12c75cfeR13) because it is not compatible with it (see [this issue](https://github.com/firebase/firebase-functions/issues/213)). However since then `firebase-functions`  has been patched, and we've upgraded so we should be able to turn the option back on.

When I tried to turn it on, the compiler detected a few places where the code was relying on some legitimate implicit anys. But it did also complain about 2 other packages that were not compatible with the option, because they were lacking types. These packages are `lodash.partition` and `firebase-tools`.

I chose to remove `lodash.partition` because I felt it was not pulling its weight, as suggested by @brunosabot  [here](https://github.com/Zenika/humeur-du-mois-2018/pull/12#issuecomment-418684338).

On the other hand, we really need `firebase-tools`, but this time instead of turning `noImplicitAny` off because of it, I took the time to study how it can be left on while using packages without types. It turns out it is possible to write your own type declarations for a module installed from npm, though it took a bit of trial and error to get there. I've documented what works in `@types/README.md`, and I've written a type declaration for the one and only function that we use from the package.