:fire: remove node version locking in function emulator scripts (#61)
bf050740ee1b69a236ea1233dda9e8cd17801037 upgraded the firebase-tools version, which broke some npm script because the path to the cli js file has changed.

I've decided to remove the dependency on the full path to the cli js file. I've reverted to using the cli as it meant to be used, without forcing the Node.js version.

I had to introduce the version locking to make the function emulator work on Windows, but it does not seem to be required anymore, it works fine without it.