⬆ make it install and run node 13 (#123)
This are the minimum (I think) changes need to make the app install and run on Node 13. In production, functions still run on Node 8.

Had to add @types/node and @type/express-serve-static-core as direct dev dependencies of functions to force the version because some packages requiring those specify * as the version so npm uses whatever version it has from other packages, which does not work because * is wrong and those package do actually need a minimum version.

Also closes #122.