📦 include auth0 in the js bundle (#160)
Close #80

Remove auth0 from the Webpack "externals" so that it is included in the js bundle.

FYI the minified auth0 script weights 148ko so the final bundle is a lot bigger than before. But I have seen that with gzip compression it falls down to ~36ko and the production app uses brotli compression so it should not be too big.