🐛 fix end of campaign reminder not sent on the right date
The number of days before the end of the campaign was not properly computed.

Running functions tests in CI, which would have revealed this bug because it was apparent when the test was run in UTC time zone. Added a test that would have shown the bug in CET time zone.